Replace DAO objects with Data Providers
Both DAO and Data Providers achieve the same result, they are just
interfaces to our data sources, it's just that I don't want to use the
DAO acronym everywhere, and Data Provider sounds more intuitive, IMHO.

I also renamed a few files, but all in all the functionality remains the
same.

Ah, and I fixed the README, the markdown for the command examples was
wrong.