Implement users controller
In this change I created a controller to handle the /users/{userId}
endpoint, where userId could be a numeric ID or a user name.

I changed the naming convention for parameters from snake_case (user_id)
to camelCase (userId) because our client is written in Javascript and it
uses camelCase.

For now our database is just a file with some users in JSON format.

This change is a needed to implement the "Has user X solved problem Y"
because for that we need users first.